Move all cluster files to cluster folder.
This will make it easier to read what belongs to cluster functionality,
as everything will be in the same folder.
