+++ redis/src/cluster/mod.rs
@@ -1,0 +1,69 @@
+//! This module extends the library to support Redis Cluster.
+//!
+//! Note that this module does not currently provide pubsub
+//! functionality.
+//!
+//! For sync functionality:
+//! # Example
+//! ```rust,no_run
+//! use redis::Commands;
+//! use redis::cluster::ClusterClient;
+//!
+//! let nodes = vec!["redis://127.0.0.1:6379/", "redis://127.0.0.1:6378/", "redis://127.0.0.1:6377/"];
+//! let client = ClusterClient::new(nodes).unwrap();
+//! let mut connection = client.get_connection().unwrap();
+//!
+//! let _: () = connection.set("test", "test_data").unwrap();
+//! let rv: String = connection.get("test").unwrap();
+//!
+//! assert_eq!(rv, "test_data");
+//! ```
+//!
+//! # Pipelining
+//! ```rust,no_run
+//! use redis::Commands;
+//! use redis::cluster::{cluster_pipe, ClusterClient};
+//!
+//! let nodes = vec!["redis://127.0.0.1:6379/", "redis://127.0.0.1:6378/", "redis://127.0.0.1:6377/"];
+//! let client = ClusterClient::new(nodes).unwrap();
+//! let mut connection = client.get_connection().unwrap();
+//!
+//! let key = "test";
+//!
+//! let _: () = cluster_pipe()
+//!     .rpush(key, "123").ignore()
+//!     .ltrim(key, -10, -1).ignore()
+//!     .expire(key, 60).ignore()
+//!     .query(&mut connection).unwrap();
+//! ```
+//!
+//! For async functionality:
+//!
+//! //! By default, [`ClusterConnection`] makes use of [`MultiplexedConnection`] and maintains a pool
+//! of connections to each node in the cluster. While it  generally behaves similarly to
+//! the sync cluster module, certain commands do not route identically, due most notably to
+//! a current lack of support for routing commands to multiple nodes.
+//!
+//! # Example
+//! ```rust,no_run
+//! use redis::cluster::ClusterClient;
+//! use redis::AsyncCommands;
+//!
+//! async fn fetch_an_integer() -> String {
+//!     let nodes = vec!["redis://127.0.0.1/"];
+//!     let client = ClusterClient::new(nodes).unwrap();
+//!     let mut connection = client.get_async_connection().await.unwrap();
+//!     let _: () = connection.set("test", "test_data").await.unwrap();
+//!     let rv: String = connection.get("test").await.unwrap();
+//!     return rv;
+//! }
+//! ```
+pub(crate) mod client;
+pub(crate) mod cluster_async;
+pub(crate) mod connections_container;
+pub(crate) mod pipeline;
+/// routing information for cluster
+pub mod routing;
+mod sync;
+pub use sync::*;
+pub(crate) mod topology;
